example for stream with user data
